software/web-app/backend/requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
sqlparse==0.5.3
tzdata==2025.1
